필터바 및 페이지네이션 버그 수정
